secure-file-sharing/backend/requirements.txt
Django==3.2.10
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
django-cors-headers==3.10.0
cryptography==3.4.7
pyotp==2.6.0